requirements.txt
google-api-python-client==2.108.0
google-auth==2.29.0
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
oauth2client==4.1.3
imaplib2==3.6
secure-smtplib==0.1.1
google-generativeai==0.4.1
SQLAlchemy==2.0.29
requests==2.31.0
slack_sdk==3.27.0
python-dateutil==2.9.0.post0
duckduckgo-search==4.5
python-dotenv==1.0.1